home_server/requirements.txt
requests==2.31.0
openai
pyserial
